requirements.txt
python-oslo.config